moved yum/dnf specific code into their plugins
